added stone; refactored
